pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebook_env"
dependencies = [
    "h5py",
    "xraylib",
    "numpy",
    "hdf5plugin",
    "swmr_tools<0.7.0",
    "blosc",
    "stomp.py",
]
dynamic = ["version"]

[project.optional-dependencies]
interactive = ["jupyterlab"]
processing = ["ipykernel", "black", "papermill"]
service = ["ipykernel", "black", "papermill_service"]
